requirements.txt
# Core dependencies
pulsectl>=23.5.2
dbus-python>=1.3.2
PyGObject>=3.46.0

# Audio processing
sounddevice>=0.4.6
webrtcvad>=2.0.10
scipy>=1.11.3
numpy>=1.24.0


